requirements.txt
minizinc
forsyde-io-python >= 0.1.2
numpy
sympy
